chore: Move polars_session.py to bigframes/testing module
